@@ -1,13 +1,13 @@
 module golang.org/x/mobile
 
-go 1.17
+go 1.18
 
 require (
-	golang.org/x/exp v0.0.0-20190731235908-ec7cb31e5a56
-	golang.org/x/image v0.0.0-20190802002840-cff245a6509b
-	golang.org/x/mod v0.6.0-dev.0.20220419223038-86c51ed26bb4
-	golang.org/x/sync v0.0.0-20220819030929-7fc1605a5dde
-	golang.org/x/tools v0.1.12
+	golang.org/x/exp/shiny v0.0.0-20230817173708-d852ddb80c63
+	golang.org/x/image v0.12.0
+	golang.org/x/mod v0.12.0
+	golang.org/x/sync v0.3.0
+	golang.org/x/tools v0.13.0
 )
 
-require golang.org/x/sys v0.0.0-20220722155257-8c9f86f7a55f // indirect
+require golang.org/x/sys v0.12.0 // indirect
